Add Terminator classes and support for AR algorithms:
Alternative routing library now support an (optional) terminator object
parameter to terminate the algorithm execution sooner if some
conditions are met.
The rationale behind this is simple: alternative routing algorithms are
very expensive and execution might take very long time on graphs with
complex topology or big number of edges. Therefore, a user might not
be anymore interested in a solution if too much time is passed. Now it
is possible to register a Terminator that will be queried periodically
to check if the execution should end.